Move RandN from torch.go into tensor.go
